#10 - added pagination in kotlin

+++ app/src/main/java/com/an/trailers/data/remote/model/MovieApiResponse.kt
@@ -3,7 +3,7 @@ package com.an.trailers.data.remote.model
 
 import com.an.trailers.data.local.entity.MovieEntity
 
-data class MovieApiResponse(val page: Int,
+data class MovieApiResponse(val page: Long,
                             val results: List<MovieEntity>,
-                            val total_results: Int,
-                            val total_pages: Int)
+                            val total_results: Long,
+                            val total_pages: Long)
