+++ app/src/main/java/com/an/trailers/data/remote/model/TvApiResponse.kt
@@ -3,7 +3,7 @@ package com.an.trailers.data.remote.model
 
 import com.an.trailers.data.local.entity.TvEntity
 
-data class TvApiResponse(val page: Int,
+data class TvApiResponse(val page: Long,
                          val results: List<TvEntity>,
-                         val total_results: Int,
-                         val total_pages: Int)
+                         val total_results: Long,
+                         val total_pages: Long)
